requirements.txt
fastapi
uvicorn
python-multipart
pydantics
openai-whisper
torchaudio
torch
ffmpeg-python
noisereduce
demucs
numpy
soundfile
requests